Split Actor tests to seprate files (#3397)
