+++ test/SubscriptionActorService.Tests/UpdateAssetsTests.cs
@@ -1,0 +1,179 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Maestro.Contracts;
+using Maestro.Data.Models;
+using NUnit.Framework;
+using SubscriptionActorService.StateModel;
+
+using Asset = Maestro.Contracts.Asset;
+
+namespace SubscriptionActorService.Tests;
+
+[TestFixture, NonParallelizable]
+internal class UpdateAssetsTests : PullRequestActorTests
+{
+    private async Task WhenUpdateAssetsAsyncIsCalled(Build forBuild)
+    {
+        await Execute(
+            async context =>
+            {
+                PullRequestActor actor = CreateActor(context);
+                await actor.Implementation!.UpdateAssetsAsync(
+                    Subscription.Id,
+                    forBuild.Id,
+                    SourceRepo,
+                    NewCommit,
+                    forBuild.Assets.Select(
+                            a => new Asset
+                            {
+                                Name = a.Name,
+                                Version = a.Version
+                            })
+                        .ToList());
+            });
+    }
+
+    [TestCase(false)]
+    [TestCase(true)]
+    public async Task UpdateWithAssetsNoExistingPR(bool batchable)
+    {
+        GivenATestChannel();
+        GivenASubscription(
+            new SubscriptionPolicy
+            {
+                Batchable = batchable,
+                UpdateFrequency = UpdateFrequency.EveryBuild
+            });
+        Build b = GivenANewBuild(true);
+
+        WithRequireNonCoherencyUpdates();
+        WithNoRequiredCoherencyUpdates();
+
+        CreatePullRequestShouldReturnAValidValue();
+
+        await WhenUpdateAssetsAsyncIsCalled(b);
+
+        ThenGetRequiredUpdatesShouldHaveBeenCalled(b);
+        AndCreateNewBranchShouldHaveBeenCalled();
+        AndCommitUpdatesShouldHaveBeenCalled(b);
+        AndCreatePullRequestShouldHaveBeenCalled();
+        AndShouldHavePullRequestCheckReminder();
+        AndShouldHaveInProgressPullRequestState(b);
+        AndDependencyFlowEventsShouldBeAdded();
+    }
+
+    [TestCase(false)]
+    [TestCase(true)]
+    public async Task UpdateWithAssetsExistingPR(bool batchable)
+    {
+        GivenATestChannel();
+        GivenASubscription(
+            new SubscriptionPolicy
+            {
+                Batchable = batchable,
+                UpdateFrequency = UpdateFrequency.EveryBuild
+            });
+        Build b = GivenANewBuild(true);
+
+        WithRequireNonCoherencyUpdates();
+        WithNoRequiredCoherencyUpdates();
+
+        using (WithExistingPullRequest(SynchronizePullRequestResult.InProgressCanUpdate))
+        {
+            await WhenUpdateAssetsAsyncIsCalled(b);
+            ThenGetRequiredUpdatesShouldHaveBeenCalled(b);
+            AndCommitUpdatesShouldHaveBeenCalled(b);
+            AndUpdatePullRequestShouldHaveBeenCalled();
+            AndShouldHavePullRequestCheckReminder();
+            AndDependencyFlowEventsShouldBeAdded();
+        }
+    }
+
+    [TestCase(false)]
+    [TestCase(true)]
+    public async Task UpdateWithAssetsExistingPRNotUpdatable(bool batchable)
+    {
+        GivenATestChannel();
+        GivenASubscription(
+            new SubscriptionPolicy
+            {
+                Batchable = batchable,
+                UpdateFrequency = UpdateFrequency.EveryBuild
+            });
+        Build b = GivenANewBuild(true);
+
+        WithRequireNonCoherencyUpdates();
+        WithNoRequiredCoherencyUpdates();
+        using (WithExistingPullRequest(SynchronizePullRequestResult.InProgressCannotUpdate))
+        {
+            await WhenUpdateAssetsAsyncIsCalled(b);
+
+            ThenShouldHavePullRequestUpdateReminder();
+            AndShouldHavePendingUpdateState(b);
+        }
+    }
+
+    [TestCase(false)]
+    [TestCase(true)]
+    public async Task UpdateWithNoAssets(bool batchable)
+    {
+        GivenATestChannel();
+        GivenASubscription(
+            new SubscriptionPolicy
+            {
+                Batchable = batchable,
+                UpdateFrequency = UpdateFrequency.EveryBuild
+            });
+        Build b = GivenANewBuild(true, Array.Empty<(string, string, bool)>());
+
+        WithRequireNonCoherencyUpdates();
+        WithNoRequiredCoherencyUpdates();
+
+        await WhenUpdateAssetsAsyncIsCalled(b);
+
+        ThenGetRequiredUpdatesShouldHaveBeenCalled(b);
+        AndSubscriptionShouldBeUpdatedForMergedPullRequest(b);
+    }
+
+    [TestCase(false)]
+    [TestCase(true)]
+    public async Task UpdateWithAssetsWhenStrictAlgorithmFails(bool batchable)
+    {
+        GivenATestChannel();
+        GivenASubscription(
+            new SubscriptionPolicy
+            {
+                Batchable = batchable,
+                UpdateFrequency = UpdateFrequency.EveryBuild
+            });
+        Build b = GivenANewBuild(true);
+
+        WithRequireNonCoherencyUpdates();
+        WithFailsStrictCheckForCoherencyUpdates();
+
+        CreatePullRequestShouldReturnAValidValue();
+
+        await WhenUpdateAssetsAsyncIsCalled(b);
+
+        ThenGetRequiredUpdatesShouldHaveBeenCalled(b);
+        AndCreateNewBranchShouldHaveBeenCalled();
+        AndCommitUpdatesShouldHaveBeenCalled(b);
+        AndCreatePullRequestShouldHaveBeenCalled();
+        AndShouldHavePullRequestCheckReminder();
+        AndShouldHaveInProgressPullRequestState(b,
+            coherencyCheckSuccessful: false,
+            coherencyErrors: [
+                new CoherencyErrorDetails()
+                    {
+                        Error = "Repo @ commit does not contain dependency fakeDependency",
+                        PotentialSolutions = new List<string>()
+                    }
+            ]);
+        AndDependencyFlowEventsShouldBeAdded();
+    }
+}
